+++ src/audio/audio.js
@@ -281,6 +281,91 @@ export function rate(sound_name, ...args) {
 }
 
 /**
+ * get or set the stereo panning for the specified sound.
+ * @function audio.stereo
+ * @param {string} sound_name - audio clip name - case sensitive
+ * @param {number} [pan] - the panning value - A value of -1.0 is all the way left and 1.0 is all the way right.
+ * @return {number} the current panning value
+ * @example
+ * me.audio.stereo("cling", -1);
+ */
+export function stereo(sound_name, pan) {
+    let sound = audioTracks[sound_name];
+    if (sound && typeof sound !== "undefined") {
+        return sound.stereo(pan);
+    } else {
+        throw new Error("audio clip " + sound_name + " does not exist");
+    }
+}
+
+/**
+ * get or set the 3D spatial position or the specified sound.
+ * @function audio.position
+ * @param {string} sound_name - audio clip name - case sensitive
+ * @param  {Number} x - the x-position of the audio source.
+ * @param  {Number} y - the y-position of the audio source.
+ * @param  {Number} z - the z-position of the audio source.
+ * @return {Array} the current 3D spatial position: [x, y, z]
+ */
+export function position(sound_name, x, y, z) {
+    let sound = audioTracks[sound_name];
+    if (sound && typeof sound !== "undefined") {
+        return sound.pos(x, y, z);
+    } else {
+        throw new Error("audio clip " + sound_name + " does not exist");
+    }
+}
+
+/**
+ * Get/set the direction the audio source is pointing in the 3D cartesian coordinate space.
+ * Depending on how direction the sound is, based on the `cone` attributes, a sound pointing away from the listener can be quiet or silent.
+ * @function audio.orientation
+ * @param {string} sound_name - audio clip name - case sensitive
+ * @param  {Number} x - the x-orientation of the audio source.
+ * @param  {Number} y - the y-orientation of the audio source.
+ * @param  {Number} z - the z-orientation of the audio source.
+ * @return {Array} the current 3D spatial orientation: [x, y, z]
+ */
+export function orientation(sound_name, x, y, z) {
+    let sound = audioTracks[sound_name];
+    if (sound && typeof sound !== "undefined") {
+        return sound.orientation(x, y, z);
+    } else {
+        throw new Error("audio clip " + sound_name + " does not exist");
+    }
+}
+
+/**
+ * get or set the panner node's attributes for a sound or group of sounds.
+ * @function audio.panner
+ * @param {string} sound_name - audio clip name - case sensitive
+ * @param {object} [attribute] - the panner attributes to set
+ * @param {string} [settings.coneInnerAngle=360] - A parameter for directional audio sources, this is an angle, in degrees, inside of which there will be no volume reduction.
+ * @param {string} [settings.coneOuterAngle=360] - A parameter for directional audio sources, this is an angle, in degrees, outside of which the volume will be reduced to a constant value of `coneOuterGain`.
+ * @param {string} [settings.coneOuterGain=0] - A parameter for directional audio sources, this is the gain outside of the `coneOuterAngle`. It is a linear value in the range `[0, 1]`.
+ * @param {string} [settings.distanceModel="inverse"] - Determines algorithm used to reduce volume as audio moves away from listener. Can be `linear`, `inverse` or `exponential.
+ * @param {string} [settings.maxDistance=10000] - The maximum distance between source and listener, after which the volume will not be reduced any further.
+ * @param {string} [settings.refDistance=1] - A reference distance for reducing volume as source moves further from the listener. This is simply a variable of the distance model and has a different effect depending on which model is used and the scale of your coordinates. Generally, volume will be equal to 1 at this distance.
+ * @param {string} [settings.rolloffFactor=1] - How quickly the volume reduces as source moves from listener. This is simply a variable of the distance model and can be in the range of `[0, 1]` with `linear` and `[0, ∞]` with `inverse` and `exponential`.
+ * @param {string} [settings.panningModel="HRTF"] - Determines which spatialization algorithm is used to position audio. Can be `HRTF` or `equalpower`.
+ * @return {Object} current panner attributes.
+ * me.audio.panner("cling", {
+ *    panningModel: 'HRTF',
+ *    refDistance: 0.8,
+ *    rolloffFactor: 2.5,
+ *    distanceModel: 'exponential'
+ * });
+ */
+export function panner(sound_name, attributes) {
+    let sound = audioTracks[sound_name];
+    if (sound && typeof sound !== "undefined") {
+        return sound.pannerAttr(attributes);
+    } else {
+        throw new Error("audio clip " + sound_name + " does not exist");
+    }
+}
+
+/**
  * stop the specified sound on all channels
  * @function audio.stop
  * @param {string} [sound_name] - audio clip name (case sensitive). If none is passed, all sounds are stopped.